% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond.effect.R
\name{cond.effect}
\alias{cond.effect}
\title{PROCESS (v3.0) Model 1 Conditional Effects and Plots}
\usage{
cond.effect(mod, Y, X, W, CV, data, se.HC, ci, digits, x.type = c("bin",
  "multicat", "cont"), w.type = c("bin", "multicat", "cont"),
  cond.type = "perc", plot.int = FALSE, plot.JN = FALSE, JN.req = FALSE)
}
\arguments{
\item{mod}{Linear model of class \code{"lm"}.}

\item{Y}{Dependent variable.}

\item{X}{Independent variable.}

\item{W}{Moderating variable.}

\item{CV}{Covariates. Defaults to \code{NULL}.}

\item{se.HC}{Use HC consistent standard errors. Options are: \code{c("none", "HC0", "HC1", "HC2", "HC3", "HC4")}. Defaults to \code{"none"}.}

\item{digits}{Number of digits to output. Defaults to \code{4}.}

\item{x.type}{Format of \code{X} variable: binary, multicategorical, or continuous. Corresponding options are \code{c("bin", "multicat", "cont")}.}

\item{w.type}{Format of \code{W} variable: binary, multicategorical, or continuous. Corresponding options are \code{c("bin", "multicat", "cont")}.#' @param data Name of data.frame or tibble.}

\item{cond.type}{Conditioning values for \code{W}. Options are \code{"perc"} for percentiles (\code{.16, .50, .84}) or \code{"mean"} for \code{M, M +/- SD}. Defaults to \code{"perc"}.}

\item{plot.int}{Plots the interaction plot. Defaults to \code{FALSE}.}

\item{plot.JN}{Plots the Johnson-Neyman areas of statistical significance. Defaults to \code{FALSE}.}

\item{JN.req}{Calculates the Johnson-Neyman areas of statistical significance. Defaults to \code{FALSE}.}

\item{p.probe}{Maximum \emph{p}-value of interaction term at which conditional effects should be probed. Defaults to \code{.10}.}
}
\description{
Implementation of the PROCESS (v3.0) Model 1 conditional effects based on Hayes (2018).
Automatically generates interaction plots and Johnson-Neyman plots of areas of significance.
Predictor or moderating variables can be continuous, binary, or multicategorical. However, both cannot be multicategorical.
}
\references{
Hayes, A. F. (2018). \emph{Introduction to Mediation, Moderation, and Conditional Process Analysis} (2e). The Guilford Press: New York.
}
\author{
Sam Mancuso
}
\keyword{1.}
\keyword{Conditional}
\keyword{Effects,}
\keyword{Model}
\keyword{Moderation,}
\keyword{PROCESS}
\keyword{macro,}
