% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model1.R
\name{model1}
\alias{model1}
\title{PROCESS (v3.0) Model 1}
\usage{
model1(Y, X, W, CV = NULL, ci = 0.95, B = 5000, boot.inf = FALSE,
  seed = 1234, digits = 4, probe.p = 0.1, mean.center = FALSE,
  se.HC = c("none", "HC0", "HC1", "HC2", "HC3", "HC4"), x.multicat = FALSE,
  x.code = c("indicator", "sequential", "helmert", "effect"),
  w.multicat = FALSE, w.code = c("indicator", "sequential", "helmert",
  "effect"), cond.type = c("perc", "mean"), JN = FALSE, plot.JN = FALSE,
  plot.int = FALSE, data)
}
\arguments{
\item{Y}{Dependent variable.}

\item{X}{Independent variable.}

\item{W}{Moderating variable.}

\item{CV}{Covariates. Defaults to \code{NULL}.}

\item{ci}{Confidence interval range. Defaults to \code{0.95}.}

\item{B}{Number of bootstrap resamples. Defaults to \code{5000}.}

\item{boot.inf}{Obtains bootstrapped confidence intervals for regression. Defaults to \code{FALSE}.}

\item{seed}{Seed value to replicate analyses. Defaults to \code{1234}.}

\item{digits}{Number of digits to output. Defaults to \code{4}.}

\item{mean.center}{Mean centre the predictors \code{X} and \code{W}. Defaults to \code{FALSE}.}

\item{se.HC}{Use HC consistent standard errors. Options are: \code{c("none", "HC0", "HC1", "HC2", "HC3", "HC4")}. Defaults to \code{"none"}.}

\item{x.multicat}{Logical value to identify if \code{X} is a multicategorical variable. Defaults to \code{FALSE}.}

\item{x.code}{Coding format for multicategorical \code{X}. Options are \code{c("indicator", "sequential", "helmert", "effect")}. Defaults to \code{"indicator"}.}

\item{w.multicat}{Logical value to identify if \code{W} is a multicategorical variable. Defaults to \code{FALSE}.}

\item{w.code}{Coding format for multicategorical \code{W}. Options are \code{c("indicator", "sequential", "helmert", "effect")}. Defaults to \code{"indicator"}.}

\item{cond.type}{Conditioning values for \code{W}. Options are \code{"perc"} for percentiles (\code{.16, .50, .84}) or \code{"mean"} for \code{M, M +/- SD}. Defaults to \code{"perc"}.}

\item{JN}{Calculates the Johnson-Neyman areas of statistical significance. Defaults to \code{FALSE}.}

\item{plot.JN}{Plots the Johnson-Neyman areas of statistical significance. Defaults to \code{FALSE}.}

\item{plot.int}{Plots the interaction plot. Defaults to \code{FALSE}.}

\item{data}{Name of data.frame or tibble.}

\item{p.probe}{Maximum \emph{p}-value of interaction term at which conditional effects should be probed. Defaults to \code{.10}.}
}
\description{
Implementation of the PROCESS (v3.0) Model 1 for moderation and conditional effects based on Hayes (2018).
Automatically generates interaction plots and Johnson-Neyman plots of areas of significance.
Predictor or moderating variables can be continuous, binary, or multicategorical. However, both cannot be multicategorical.
}
\references{
Hayes, A. F. (2018). \emph{Introduction to Mediation, Moderation, and Conditional Process Analysis} (2e). The Guilford Press: New York.
}
\author{
Sam Mancuso
}
\keyword{1.}
\keyword{Conditional}
\keyword{Effects,}
\keyword{Model}
\keyword{Moderation,}
\keyword{PROCESS}
\keyword{macro,}
